refactor(task): replace magic strings with Type enum
