+++ src/main/java/br/com/alura/AluraFake/task/TaskRepository.java
@@ -2,6 +2,7 @@ package br.com.alura.AluraFake.task;
 
 import br.com.alura.AluraFake.course.Course;
 import br.com.alura.AluraFake.task.model.Task;
+import br.com.alura.AluraFake.task.model.Type;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -16,8 +17,8 @@ public interface TaskRepository extends JpaRepository<Task, Long> {
 
     List<Task> findByCourseAndOrderGreaterThanEqualOrderByOrderAsc(Course course, Integer order);
 
-    @Query("SELECT COUNT(t) FROM Task t WHERE t.course = :course AND t.taskType = :taskTypeString")
-    long countByCourseAndTaskType(@Param("course") Course course, @Param("taskTypeString") String taskTypeString);
+    @Query("SELECT COUNT(t) FROM Task t WHERE t.course = :course AND t.taskType = :#{#taskType.name()}")
+    long countByCourseAndTaskType(@Param("course") Course course, @Param("taskType") Type taskType);
 
     @Query("SELECT COUNT(t) > 0 FROM Task t WHERE t.course = :course AND t.order = :order")
     boolean existsByCourseAndOrder(@Param("course") Course course, @Param("order") int order);
